Update gitignore and test-requirements
After the last module sync, there were a few items that need to be
updated. The gitignore needs /openstack rather than openstack/ to ensure
that we're only ignoreing the base /openstack folder that gets created
when you run the tests. Additionall the test-requirements.txt was
updated with a newer list of requirements for the release notes.

Change-Id: Id057d8d947337e501f0dd41cf7694d56d26b1c25